refactor(core): extract TreeNode to WidgetTreeNode
Move TreeNode to new WidgetTreeNode class for clarity.
Update usages to use WidgetTreeNode, improving code organization.
