+++ Ivy/Core/WidgetTree.cs
@@ -12,103 +12,6 @@ using Ivy.Core.Hooks;
 
 namespace Ivy.Core;
 
-public class TreeNode(string id, int index, TreePath parentTreePath, TreeNode[] children, int? memoizedHashCode, IWidget? widget, IView? view, IViewContext? context, IViewContext? ancestorContext) : IDisposable
-{
-    public static TreeNode FromWidget(IWidget widget, int index, TreePath treePath, TreeNode[] children, IViewContext? ancestorContext)
-    {
-        return new TreeNode(widget.Id!, index, treePath, children, null, widget, null, null, ancestorContext);
-    }
-
-    public static TreeNode FromView(IView view, int index, TreePath treePath, TreeNode? child, IViewContext? context, int? memoizedHashCode, IViewContext? ancestorContext)
-    {
-        var children = child == null ? Array.Empty<TreeNode>() : [child];
-        return new TreeNode(view.Id!, index, treePath, children, memoizedHashCode, null, view, context, ancestorContext);
-    }
-
-    // public bool ShouldRebuild { get; set; }
-
-    public string Id { get; } = id;
-    public int Index { get; } = index;
-    public TreePath ParentTreePath { get; } = parentTreePath;
-    public TreeNode[] Children { get; } = children;
-    public int? MemoizedHashCode { get; } = memoizedHashCode;
-    public IWidget? Widget { get; } = widget;
-    public IView? View { get; } = view;
-    public IViewContext? Context { get; } = context;
-    public IViewContext? AncestorContext { get; } = ancestorContext;
-    public Type NodeType => IsWidget ? Widget!.GetType() : View!.GetType();
-
-    public IWidget? GetWidgetTree()
-    {
-        if (IsWidget)
-        {
-            var children = Children
-                .Select(child => child.GetWidgetTree()).Cast<object>().ToArray();
-
-            if (Widget is AbstractWidget widget)
-            {
-                return widget with
-                {
-                    Children = children
-                };
-            }
-
-            throw new NotSupportedException("Widgets must be of type WidgetBase.");
-        }
-
-        if (IsView)
-        {
-            //views always have none or one child
-            return Children.SingleOrDefault()?.GetWidgetTree();
-        }
-
-        throw new NotSupportedException("Node must be either an IWidget or an IView.");
-    }
-
-    public bool IsWidget => Widget != null;
-    public bool IsView => View != null;
-
-    public void Dispose()
-    {
-        View?.Dispose();
-        Context?.Dispose();
-    }
-
-    public int[] GetWidgetTreeIndices()
-    {
-        //The top of the tree is always a view
-        //A view can only have one child which is either a view or a widget
-        //We want get the indices as if the views has been removed and only the widgets remain
-
-        var indices = new List<int>() { };
-        var path = this.ParentTreePath.Clone();
-
-        var previousSegment = new PathSegment("", null, this.Index, this.IsWidget);
-
-        while (path.Count > 0)
-        {
-            var segment = path.Pop();
-            if (segment.IsWidget)
-            {
-                indices.Add(previousSegment.Index);
-            }
-            previousSegment = segment;
-        }
-
-        indices.Reverse();
-        return indices.ToArray();
-    }
-
-    // public void CancelRebuild()
-    // {
-    //     this.ShouldRebuild = false;
-    //     foreach (var child in Children)
-    //     {
-    //         child.CancelRebuild();
-    //     }
-    // }
-}
-
 public class WidgetTreeChanged(string viewId, int[] indices, JsonNode patch, int iteration, string? treeHash)
 {
     public string ViewId { get; } = viewId;
@@ -120,15 +23,15 @@ public class WidgetTreeChanged(string viewId, int[] indices, JsonNode patch, int
 
 public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
 {
-    private readonly Dictionary<string, TreeNode> _nodes = new();
+    private readonly Dictionary<string, WidgetTreeNode> _nodes = new();
     private readonly Dictionary<string, string> _parents = new();
-    private int _iteration = 0;
+    private int _iteration;
 
     public IView RootView { get; }
-    public TreeNode? NodeTree { get; private set; }
+    public WidgetTreeNode? NodeTree { get; private set; }
 
-    public static JsonPatchDeltaFormatter JsonPatchDeltaFormatter = new();
-    public static JsonDiffOptions JsonDiffOptions = new()
+    public static readonly JsonPatchDeltaFormatter JsonPatchDeltaFormatter = new();
+    public static readonly JsonDiffOptions JsonDiffOptions = new()
     {
         ArrayObjectItemKeyFinder = (node, _) =>
         {
@@ -307,8 +210,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
 
         var indices = node.GetWidgetTreeIndices();
 
-        //todo: We are serializing twice. Is it worth caching the previous serialized tree in the node?
-        var previous = node.GetWidgetTree()?.Serialize();
+        var previous = node.GetSerializedWidgetTree();
 
         var partial = BuildView(node.View!, node.ParentTreePath.Clone(), node.Index, parentId, node.AncestorContext, isRefreshingView: true, isHotReload);
 
@@ -339,7 +241,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
 
         try
         {
-            var update = partial.GetWidgetTree()?.Serialize();
+            var update = partial.GetSerializedWidgetTree();
 
             var previousId = previous?["id"]?.GetValue<string>();
             var updateId = update?["id"]?.GetValue<string>();
@@ -386,7 +288,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
         return null!;
     }
 
-    private TreeNode? BuildView(IView view,
+    private WidgetTreeNode? BuildView(IView view,
         TreePath treePath,
         int index,
         string? parentId,
@@ -410,7 +312,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
 
         var previousNode = _nodes.GetValueOrDefault(view.Id);
 
-        TreeNode? node;
+        WidgetTreeNode? node;
         IViewContext? context = previousNode?.Context;
         if (!memoized || isHotReload || isRefreshingView || previousNode == null || previousNode.MemoizedHashCode != memoizedHashCode)
         {
@@ -469,7 +371,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
 
         treePath.Pop();
 
-        _nodes[view.Id] = TreeNode.FromView(view, index, treePath.Clone(), node, context, memoizedHashCode, ancestorContext);
+        _nodes[view.Id] = WidgetTreeNode.FromView(view, index, treePath.Clone(), node, context, memoizedHashCode, ancestorContext);
 
         if (parentId != null)
             _parents[view.Id] = parentId;
@@ -514,7 +416,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
         return hash.ToHashCode();
     }
 
-    private TreeNode? BuildObject(object? anything, TreePath treePath, int index, string parentId, IViewContext? ancestorContext, bool isHotReload)
+    private WidgetTreeNode? BuildObject(object? anything, TreePath treePath, int index, string parentId, IViewContext? ancestorContext, bool isHotReload)
     {
         var formatted = _contentBuilder.Format(anything);
         if (formatted == null) return null;
@@ -522,7 +424,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
         if (formatted is not IView && formatted is not IWidget)
             throw new NotSupportedException("IContentFormatter must return either an IView or an IWidget.");
 
-        TreeNode? node = null;
+        WidgetTreeNode? node = null;
         if (formatted is IView newView)
         {
             node = BuildView(newView, treePath.Clone(), index, parentId, ancestorContext, false, isHotReload);
@@ -536,13 +438,13 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
         return node;
     }
 
-    private TreeNode BuildWidget(IWidget widget, TreePath treePath, int index, string parentId, IViewContext? ancestorContext, bool isHotReload)
+    private WidgetTreeNode BuildWidget(IWidget widget, TreePath treePath, int index, string parentId, IViewContext? ancestorContext, bool isHotReload)
     {
         treePath.Push(widget, index);
 
         widget.Id = treePath.GenerateId();
 
-        var children = new List<TreeNode>();
+        var children = new List<WidgetTreeNode>();
         if (widget.Children == null!) widget.Children = [];
         for (var i = 0; i < widget.Children.Length; i++)
         {
@@ -554,7 +456,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
 
         treePath.Pop();
 
-        var node = TreeNode.FromWidget(widget, index, treePath.Clone(), children.ToArray(), ancestorContext);
+        var node = WidgetTreeNode.FromWidget(widget, index, treePath.Clone(), children.ToArray(), ancestorContext);
 
         _nodes[widget.Id] = node;
         _parents[widget.Id] = parentId;
@@ -592,7 +494,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
         //PrintTree(NodeTree, 0);
     }
 
-    private void PrintTree(TreeNode? node, int i)
+    private void PrintTree(WidgetTreeNode? node, int i)
     {
         if (node == null) return;
         Console.Write(new string(' ', i));
@@ -624,7 +526,7 @@ public class WidgetTree : IWidgetTree, IObservable<WidgetTreeChanged[]>
         }
     }
 
-    private void DestroyRemovedNodes(TreeNode previousNode, TreeNode node, string? skipViewId)
+    private void DestroyRemovedNodes(WidgetTreeNode previousNode, WidgetTreeNode node, string? skipViewId)
     {
         if (previousNode.Id != node.Id)
         {
